# Make rendering test faster using updated yamatanooroti