Makes test_rendering faster. Yamatanooroti's wait time is reduced from 0.1 sec to 0.01 sec
`test_yamatanooroti(vterm-yamatanooroti head)`: 37.58 sec → 10.72 sec


## startup_message
We don't need `puts 'start IRB'`. It does not mean IRB is ready for input. We should check the prompt with `startup_message: /irb\(main\)/`

## close
It does not only close_input but read few bytes and terminates process immediately.
Calling close before assert_screen sometimes terminates process before reading the last output.